+++ lib/baabedo/channel.rb
@@ -1,0 +1,5 @@
+module Baabedo
+  class Channel < APIResource
+    include Baabedo::APIOperations::List
+  end
+end