+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Next (?/?/?)
 #### Features
 
 * Added `proxy` option to allow RestClient http proxies.
+* Added Baabedo::Order. (list/search is still missing)
 
 0.0.4 (23/08/2015)
 ==================